MAJOR CORE UPDATE
Complete rework of the core files. New input-output structure and organization of  the already written Cpp code.
feedback, setpoint_mannager and input_mannger is almost entirely re-written. Added several helper files for cleaning up the core files (*_gen). The intent is to absorb as much of the repetitious elements into separate classes and develop template code for each of the managers.

This update completely changed variable and class field names of the setpoint_manager and input manager.

Note. Feedforward is not actually being written anywhere yet.